requirements.txt
requests
beautifulsoup4
serpapi
rank-bm25
groq
python-dotenv
tqdm